refactor(httpapi): move openapi normalization out of main
